pyproject.toml
[build-system]
requires = ["setuptools>=77.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "uassistant"
version = "2.0.0"
description = "Private, multilingual, offline AI assistant"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
dependencies = [
    "customtkinter>=5.2.0",
    "pillow>=10.0.0",
    "sounddevice>=0.4.6",
    "faster-whisper>=1.0.1",
    "sentence-transformers>=2.2.0",
    "numpy>=1.26.4",
    "PyPDF2>=3.0.1",
    "openpyxl>=3.1.2",
    "python-docx>=1.1.2",
    "requests>=2.32.3"
]

[project.scripts]
uassistant = "uassistant.main:main"

[tool.setuptools.packages.find]
where = ["src"]